+++ src/reducers/usage.js
@@ -1,13 +1,14 @@
 import {
 	RECEIVE_USAGE,
 	LOGIN_ERROR,
-	RECEIVE_LAB_USAGE
+	RECEIVE_LAB_USAGE,
+	RECEIVE_ALLOCATIONS
 } from "../actions";
 
 const defaultState = {
 	usage: {},
-	hazardUsage: {},
-	labUsage: {}
+	labUsage: {},
+	allocations: {}
 };
 
 const usage = (state=defaultState, action) => {
@@ -16,6 +17,8 @@ const usage = (state=defaultState, action) => {
 		return Object.assign({}, state, {usage: action.usage});
 	case RECEIVE_LAB_USAGE:
 		return Object.assign({}, state, {lab: action.usage});
+	case RECEIVE_ALLOCATIONS:
+		return Object.assign({}, state, {allocations: action.allocations});
 	case LOGIN_ERROR:
 		return Object.assign({}, state, {datasets: [], Authorization: "", loginError: true});
 	default: